% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{This function collects comments from source code that use the #' symbol for commenting.
It also collects lines of code}
\usage{
get_comments(file, encoding = "UTF-8")
}
\arguments{
\item{file}{A string that contains the path to the source file}

\item{encoding}{A string containing the encode that will be used to read the script, by default it will be 'Native'}
}
\value{
a list of lines
}
\description{
This function collects comments from source code that use the #' symbol for commenting.
It also collects lines of code
}
