% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_md.R
\name{make_md}
\alias{make_md}
\title{Parses a list of comments in markdown style and
renders a markdown or html file.
To be used in conjunction with \code{get_comments()}}
\arguments{
\item{file}{A string containing the path of the file that will be documented}

\item{output_name}{A string containing the name of the file that will be written, without extension}

\item{encoding}{A string containing the encode that will be passed on to the \code{get_comments()} and \code{markdownToHTML} functions, by default it will be 'Native'}

\item{style}{A string containing a CSS stylesheet file}

\item{...}{args that can be passed on to markdownToHTML}
}
\value{
A markdown and a html file.
}
\description{
Parses a list of comments in markdown style and
renders a markdown or html file.
To be used in conjunction with \code{get_comments()}
}
